Move issue-creator to robots directory.
