+++ BUILD
@@ -32,7 +32,6 @@ filegroup(
         "//gcsweb/pkg/version:all-srcs",
         "//ghclient:all-srcs",
         "//images/pull_kubernetes_bazel:all-srcs",
-        "//issue-creator:all-srcs",
         "//jenkins:all-srcs",
         "//jobs:all-srcs",
         "//kettle:all-srcs",
@@ -42,6 +41,7 @@ filegroup(
         "//metrics:all-srcs",
         "//mungegithub:all-srcs",
         "//prow:all-srcs",
+        "//robots/issue-creator:all-srcs",
         "//scenarios:all-srcs",
         "//testgrid/config:all-srcs",
         "//testgrid/jenkins_verify:all-srcs",
