+++ robots/issue-creator/sources/BUILD
@@ -1,0 +1,44 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "flakyjob-reporter.go",
+        "triage-filer.go",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//mungegithub/mungers/mungerutil:go_default_library",
+        "//robots/issue-creator/creator:go_default_library",
+        "//vendor/github.com/golang/glog:go_default_library",
+        "//vendor/github.com/google/go-github/github:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "flakyjob-reporter_test.go",
+        "triage-filer_test.go",
+    ],
+    library = ":go_default_library",
+    deps = [
+        "//robots/issue-creator/creator:go_default_library",
+        "//robots/issue-creator/testowner:go_default_library",
+        "//vendor/github.com/google/go-github/github:go_default_library",
+    ],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)
