+++ robots/issue-creator/sources/triage-filer.go
@@ -28,8 +28,8 @@ import (
 	"time"
 
 	githubapi "github.com/google/go-github/github"
-	"k8s.io/test-infra/issue-creator/creator"
 	"k8s.io/test-infra/mungegithub/mungers/mungerutil"
+	"k8s.io/test-infra/robots/issue-creator/creator"
 )
 
 const (
