+++ robots/issue-creator/sources/triage-filer_test.go
@@ -24,8 +24,8 @@ import (
 	"time"
 
 	"github.com/google/go-github/github"
-	"k8s.io/test-infra/issue-creator/creator"
-	"k8s.io/test-infra/issue-creator/testowner"
+	"k8s.io/test-infra/robots/issue-creator/creator"
+	"k8s.io/test-infra/robots/issue-creator/testowner"
 )
 
 var (
